product_agent/agent.py
from google.adk.agents.llm_agent import Agent
import os
from .tools.my_tools import *
from shared_tools import *

product_agent = Agent(
    name="ProductSupport",
    model=os.getenv("ADK_MODEL_NAME"),
    description="Handles product information, pricing, feature comparison, and sales questions",
    instruction="""You are a product specialist. You handle:
- Product information and capabilities
- Pricing and plan comparisons
- Feature explanations
- Use case questions
- Upgrade/downgrade recommendations

When helping customers:
1. Understand their needs and use case
2. Look up pricing and feature availability
3. Recommend appropriate plans
4. Explain features clearly
5. Use the knowledge base for detailed info

Help customers find the right solution for their needs.

IMPORTANT: When you receive a request, immediately:
- Acknowledge their product question
- Use get_pricing_info to provide current pricing
- Use check_feature_availability for feature questions
- Recommend the best plan for their use case
- Provide clear benefits and next steps

Always provide a concrete response with actionable information.

Once the request is completed, send the message back to the coordinator to relay to the customer.""",
    tools=[
        get_pricing_info,
        check_feature_availability,
        search_knowledge_base,
        create_support_ticket
    ]
)

root_agent = product_agent
