requirements.txt
google_adk
litellm>=1.75.5